requirements.txt
django==3.2
python-decouple
pillow
djangorestframework
markdown
django-filter
coreapi
drf-yasg[validation]
djangorestframework-simplejwt
django-mail-templated
djangorestframework

black
flake8
pytest
pytest-django